deepen: absorb derive-stake into form4-field-normalizer
Inlines lib/ai/utils/derive-stake.ts (234 LOC, 2 callers) into
lib/email/form4-field-normalizer.ts, its primary caller. Also updates
the second caller (lib/ai/parsers/response-parser.ts) to import
deriveNewStake from form4-field-normalizer alongside the other Form 4
primitives it already draws from that module.

Motivation: the deleted file lived in a documented circular import
with form4-field-normalizer.ts. Its own header comment warned that
the cycle was 'safe only because formatNumberWithCommas is a function
declaration and gets hoisted. Do NOT convert it to const fn = () => ...
or the cycle crashes at module init.' That is exactly the load-order
fragility the deepening rules ('one adapter is a hypothetical seam')
warn against — the extracted module existed for testability, not
locality, and its stake-derivation logic conceptually lives with the
NormalizedTransaction shape it operates on.

Zero tests deleted; zero tests added. derive-stake.ts had no test
file of its own. Coverage is exercised through form4-field-normalizer
(the seam that now owns the code) and response-parser-normalization
tests. All 1135 tests across __tests__/ai/parsers + __tests__/email
pass locally.

CONTEXT.md: new 'Post-Transaction Stake' term documents the
five-tier derivation and the deepening.
