+++ CONTEXT.md
@@ -84,5 +84,8 @@ Production code reaches SEC EDGAR through a small, hand-picked surface in `lib/s
 ### Monitoring
 Operational counters, gauges, and timers recorded across the cron pipeline, AI clients, and email send paths. Implemented as one deep singleton `monitoring` in `lib/monitoring/index.ts` exposing `incrementCounter` / `recordValue` / `recordGauge` / `startTimer` / `stopTimer` / `recordTiming` / `recordHealth` / `recordAiApiCall`, plus a per-metric PostHog forwarding bridge gated on the `POSTHOG_METRIC_PREFIX_ALLOWLIST` (`ai.x_sentiment_`, `ai.ticker_grounding_`, `ai.why_it_matters_`, `ai.numeric_grounding_`). Two adjacent files keep their own interfaces because they have their own production callers: `cron-monitor.ts` (`CronJobMonitor`, used by `app/api/cron/route.ts`) and `json-parsing-monitor.ts` (`jsonParsingMonitor`, used by `lib/ai/parsers/response-parser.ts`). `cron-monitor.ts` privately consumes `async-alert-queue.ts` and `performance-monitor.ts`, which are internal to the cron-alert pipeline and never reach external callers. Replaces an orbital cluster of zero-production-caller modules that the `lib/monitoring/index.ts` barrel had been re-exporting: `pipeline-error-detector.ts` (~999 LOC `PipelineErrorDetector` singleton with `ErrorPattern` / `ErrorAlert` / `PipelineMetrics` / `DetectionReport` taxonomy and a `detectPipelineErrors` / `generatePipelineHealthReport` / `resolvePipelineAlert` surface), `pipeline-health-monitoring-system.ts` (~1040 LOC `PipelineHealthMonitoringSystem extends EventEmitter` singleton with `HealthMonitoringConfig` / `HealthMetrics` / `AlertEvent` / `DashboardMetrics` taxonomy plus `recordLockOperation` / `recordParameterValidation` / `recordSecApiOperation` / `recordUserProcessing` / `getCurrentHealthMetrics` / `getDashboardMetrics` / `startPipelineHealthMonitoring` / `stopPipelineHealthMonitoring` entry points), and `sec-api-monitor.ts` (~607 LOC `MonitoredSECEdgarClient extends SECEdgarClient` adapter plus a `SecApiMonitor` class, exporting `monitoredSecClient` as the canonical instance). All three were only ever consumed by each other (pipeline-health → pipelineErrorDetector, sec-api-monitor → recordSecApiOperation) and re-exported through the barrel; every external `from '@/lib/monitoring'` import only ever destructures `{ monitoring }`. Deletion collapses a 3-candidate hypothetical seam (3 adapters under the barrel, 0 callers) to the singleton surface production actually exercises.
 
+### Post-Transaction Stake
+The `newStake` field on a Form 3/4/144 [Summary] — the insider's holdings AFTER the reported transaction, deterministically derived from the LLM's `transactions[]` array rather than trusted from the LLM's top-level `newStake` string. Five-tier precedence: (1) authoritative LLM `postTransactionCommonShares`, (2) common-stock direct-ownership `sharesOwnedFollowing` sorted chronologically, (3) fallback to any transaction with `sharesOwnedFollowing`, (4) LLM's legacy top-level string, (5) narrative regex over `summaryText`. Background: the LLM's `newStake` was silently pulled from Table II (derivative) rows in the AAPL Parekh 2026-04-15 filing when the Table I Column 5 Common Stock row was the correct source; the deterministic derivation ships the correct number regardless of which row the LLM picked. Implemented as `deriveNewStake` + `detectNewStakeNarrativeMismatch` inside `lib/email/form4-field-normalizer.ts` — not a separate module — because the derivation depends on the same `NormalizedTransaction` shape and `formatNumberWithCommas` helper as the rest of Form 4 normalization. Replaces a former `lib/ai/utils/derive-stake.ts` that had two callers (form4-field-normalizer.ts and `lib/ai/parsers/response-parser.ts`) but sat in a documented circular import with form4-field-normalizer.ts, whose only reason for existing was load-order fragility ("Safe only because formatNumberWithCommas is a `function` declaration and gets hoisted. Do NOT convert it to `const fn = () => ...` or the cycle crashes at module init"). Merging concentrates all Form 4 stake locality in one file and eliminates the fragile hoisting dependency. `response-parser.ts` now imports `deriveNewStake` alongside the other Form 4 primitives it already draws from this module.
+
 ### Resilience
 Retry-with-backoff and circuit-breaker behaviour applied at every outbound call seam (OpenRouter client, Anthropic client, Resend email, SEC fetch, network helper). Implemented as one deep module pair behind `executeWithRetry` + `executeWithAdaptiveRetry`: `lib/error-handling/retry.ts` carries the fixed/exponential retry loop and the `CircuitBreaker` class; `lib/error-handling/adaptive-retry.ts` adds the latency-aware variant used by `enhanced-claude-client.ts` and `network/enhanced-fetch.ts`. Callers pass a function and a config; the module owns attempt counting, jitter, status-aware retry classification, breaker state machine, and observability hooks. Four parallel shallow implementations previously orbited this concept with **zero non-test callers** — `lib/resilience/circuit-breaker.ts` + `error-handling.ts` + `monitoring.ts` + `retry-utility.ts` (the `lib/resilience/` package), plus the stray `lib/utils/circuitBreaker.ts` — each carrying its own `CircuitBreakerConfig` / `RetryConfig` shape and its own state machine. Their existence forced maintainers to ask "which retry module do I import?" before every change. Deleting them collapses the hypothetical seam (4 candidate adapters, 0 callers) to one real interface, so the surviving canonical module's leverage is unambiguous. A second orbit lived under `lib/db/` — `circuit-breaker.ts`, `connection-warmer.ts`, `connection-manager.ts`, `optimized-connection-manager.ts`, `check-connections.ts`, and `monitoring.ts` (~1037 LOC) — each defining its own `CircuitBreakerConfig` / pool-stats / telemetry shape, none of them imported by production. They were deleted for the same reason: a hypothetical seam (6 candidate adapters, 0 callers) is not a seam. The lone surviving Prisma-flavoured preset `dbRetry` in `lib/db/retry-wrapper.ts` (one production caller, `app/api/user/tickers/route.ts`) keeps its Prisma-specific error-code mapping (`P1001` / `P1008` / `P1017` and `PrismaClientInitializationError`) local to the database seam — distinct enough from the generic HTTP error taxonomy in `executeWithRetry` that the duplication earns its keep. Future Resilience work — bulkheads, hedged requests, deadline propagation — extends `lib/error-handling/` rather than spawning a new sibling package.
